- Added ers handling for skill_timerskill structures.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@5995 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/guild.c
@@ -367,7 +367,7 @@ int guild_payexp_timer_sub(DBKey dataid, void *data, va_list ap)
 		return 0;
 	}
 
-	if (g->member[i].exp > INT_MAX - c->exp)
+	if ((unsigned int)g->member[i].exp > INT_MAX - c->exp)
 		g->member[i].exp = INT_MAX;
 	else
 		g->member[i].exp+= c->exp;
